add pagination on estimate list and refactor code on staff screen for both devices to work with new changes
